optimize auditlog for agent merging (#3855)
* optimize auditlog for agent merging

* query logic fix

* typo fix

* generalizable merging optimization and progress reporting

* change newrecordid column size from small int to int

* Fix serialization error in finishing message

* Make sure the finishing message is sent when serialization fails

* fix merge_status

* remove new_record_data from message + code cleanup

* fix unit test

+++ specifyweb/notifications/models.py
@@ -18,7 +18,7 @@ class Spmerging(models.Model):
     taskid = models.CharField(max_length=256) 
     mergingstatus = models.CharField(max_length=256)
     table = models.CharField(max_length=256)
-    newrecordid = models.SmallIntegerField(null=True)
+    newrecordid = models.IntegerField(null=True)
     newrecordata = models.JSONField(null=True)
     oldrecordids = models.JSONField(null=True)
     collection = models.ForeignKey(Collection, on_delete=models.CASCADE)
